refactor(davinci): fit the program docs and grown modules under the source budget
Against a main merge base the source-length gate sees davinci's early
direct-push docs as new files: prior-art.md (533) and architecture.md
(435) each split into a sibling page, davinci_harness report.rs (530)
moves its schema validator and tests into #[path] submodules, the
diagnostics runner returns under 350 by moving its test module out, and
check_cli.rs drops the one blank line it had grown by. Zero behavior
change; every gate that reads these files passes with origin/main as the
base ref.
