+++ davinci-road/README.md
@@ -88,17 +88,17 @@ Recorded 2026-08-13 after design review. Revisit requires a written entry in
 
 ## Documents
 
-| Document                                | Contents                                                                                                                                                                                                                                  |
-| --------------------------------------- | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| [Motivation](./motivation.md)           | Current-state fault lines with file-path evidence, and the existing assets Davinci builds on                                                                                                                                              |
-| [Architecture](./architecture.md)       | The stage model (S0–S4), stage fusion, shared infrastructure, dialect and target contracts, performance guardrails                                                                                                                        |
-| [Semantic Engine](./semantic-engine.md) | The analyzer pillar: measured Croquis underuse, the fact/query design, the reactivity lattice serving Vapor and non-Vapor alike, app-level facts, complexity and HTML-conformance products                                                |
-| [DevTool](./devtool.md)                 | The observability surface: stage ladder, pass timeline, provenance, fact browser, decision remarks, flame views                                                                                                                           |
-| [Assurance](./assurance.md)             | The quality doctrine: impossibility by construction, input-space enumeration, the test tier ladder, strict oracles, mutation-tested tests                                                                                                 |
-| [Roadmap](./roadmap.md)                 | Phases, exit gates, and risks                                                                                                                                                                                                             |
-| [Prior Art](./prior-art.md)             | Practices imported from rustc/MIR/Polonius/salsa, LLVM/MLIR, Swift (SwiftSyntax/SIL/macros), GHC (Core Lint/interface files), OCaml Flambda2, Lean 4, React Compiler, MoonBit, Unison, Effekt, and recent PL research — with anti-lessons |
-| [Implementation Plan](./plan/README.md) | PR-granular task decomposition: phases 0–1 in full detail with Steps, phases 2–6 provisional, the [test-suite registry](./plan/test-suites.md) (TS-1..51), and continuous workstreams                                                     |
-| [Open Questions](./open-questions.md)   | Active design discussions not yet decided                                                                                                                                                                                                 |
+| Document                                | Contents                                                                                                                                                                                                                                                                                      |
+| --------------------------------------- | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| [Motivation](./motivation.md)           | Current-state fault lines with file-path evidence, and the existing assets Davinci builds on                                                                                                                                                                                                  |
+| [Architecture](./architecture.md)       | The stage model (S0–S4), stage fusion, priority order, performance guardrails; [shared infrastructure and the extension contracts](./architecture-infrastructure.md) are split out under the source budget                                                                                    |
+| [Semantic Engine](./semantic-engine.md) | The analyzer pillar: measured Croquis underuse, the fact/query design, the reactivity lattice serving Vapor and non-Vapor alike, app-level facts, complexity and HTML-conformance products                                                                                                    |
+| [DevTool](./devtool.md)                 | The observability surface: stage ladder, pass timeline, provenance, fact browser, decision remarks, flame views                                                                                                                                                                               |
+| [Assurance](./assurance.md)             | The quality doctrine: impossibility by construction, input-space enumeration, the test tier ladder, strict oracles, mutation-tested tests                                                                                                                                                     |
+| [Roadmap](./roadmap.md)                 | Phases, exit gates, and risks                                                                                                                                                                                                                                                                 |
+| [Prior Art](./prior-art.md)             | Practices imported from rustc/MIR/Polonius/salsa, LLVM/MLIR, React Compiler, MoonBit, Unison, Effekt; [the second half](./prior-art-toolchains.md) covers Swift (SwiftSyntax/SIL/macros), GHC (Core Lint/interface files), OCaml Flambda2, Lean 4, and recent PL research — with anti-lessons |
+| [Implementation Plan](./plan/README.md) | PR-granular task decomposition: phases 0–1 in full detail with Steps, phases 2–6 provisional, the [test-suite registry](./plan/test-suites.md) (TS-1..51), and continuous workstreams                                                                                                         |
+| [Open Questions](./open-questions.md)   | Active design discussions not yet decided                                                                                                                                                                                                                                                     |
 
 ## Relationship to the mission
